AgentCore/requirements.txt
mcp
strands-agents-tools
strands-agents
bedrock-agentcore
